[feat] #28 운송 가능 차량 목록 조회 api 구현

+++ module-domain/domain-transports/src/main/java/com/chaing/domain/transports/entity/Vehicle.java
@@ -46,7 +46,7 @@ public class Vehicle {
     private String driverPhone;
 
     @Column(nullable = false)
-    private Double maxLoad;
+    private Long maxLoad;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
